tools/image/requirements.txt
streamlit>=1.32.0
Pillow>=10.3.0
pdf2image>=1.17.0
regex>=2023.12.25
watchdog>=4.0.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
